R_I_FM

+++ main/method/innovation/multi_view.py
@@ -12,7 +12,7 @@ class FeatureFusion:
         c = features.size(1)
 
         integrating_features = torch.zeros([chunk_size, c]).to(features.device)
-        integrating_pids = torch.zeros([chunk_size]).to(pids.device)
+        integrating_pids = torch.zeros([chunk_size], dtype=torch.long).to(pids.device)
 
         for i in range(chunk_size):
             integrating_features[i] = 1 * (features[4 * i] + features[4 * i + 1] + features[4 * i + 2] + features[4 * i + 3])
